ez use for user to remove products add auth 2 some part of backend removed som console.log

+++ frontend/src/page/remove.js
@@ -1,26 +1,47 @@
-import e from 'cors'
+
 import React from 'react'
 import { useState } from 'react'
 import { deleteItems } from '../store/upploadingReducer'
-import { useDispatch } from 'react-redux'
+import { useDispatch, useSelector } from 'react-redux'
+import { getItem } from '../store/shopItemsReducer'
 function Remove() {
 
-    const dispatch = useDispatch()
-    const [def, setDef] = useState('')
+  const dispatch = useDispatch()
+  const [def, setDef] = useState('')
 
-   const deleteFn = (e) => { 
-    e.preventDefault()
-    dispatch(deleteItems({id: def}))
-   }
+  const item = useSelector(state => state.Entities.shop.item)
+
+  const areUShure = () => {
+    
+    let Deleteing = prompt("skriv delete sen tryck ok för att ta bort produkten", "")
+   
+    if(Deleteing.toLowerCase() === "delete") {
+      dispatch(deleteItems({id: def}))
+    }
+  }
+
+  const findProduckt = () => {
+    
+    dispatch(getItem({id: def}))
+  }
+
+
 
 
   return (
-    <div>
-        <form>
-            <input type="text" onChange={(e) => setDef(e.target.value)} />
-            <button onClick={deleteFn}>test</button>
-        </form>
-    </div>
+    
+      <div className='remove-main'>
+
+        <input className='' type="text" onChange={(e) => setDef(e.target.value)}/>
+        
+        {item && <img className='img-remove' src={`data:image/png;base64,${item.bild}`} alt="bild"/>}
+        
+        <div className='btns-remove'>
+          <button onClick={findProduckt} >Find</button>
+          <button onClick={areUShure} >Delete</button>
+        </div>
+
+      </div>
   )
 }
